Refs #19049. Move peaks statistics stuff into separate file
